extact methods from IpcamSnapshotActivity into MjpegRecordingHandler
