+++ app/src/main/java/com/github/niqdev/ipcam/IpCamSnapshotActivity.java
@@ -15,12 +15,10 @@ import android.widget.Toast;
 import com.github.niqdev.mjpeg.DisplayMode;
 import com.github.niqdev.mjpeg.Mjpeg;
 import com.github.niqdev.mjpeg.MjpegView;
-import com.github.niqdev.mjpeg.OnFrameCapturedListener;
 
 import java.io.BufferedOutputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.text.SimpleDateFormat;
@@ -37,11 +35,10 @@ import static com.github.niqdev.ipcam.settings.SettingsActivity.PREF_AUTH_PASSWO
 import static com.github.niqdev.ipcam.settings.SettingsActivity.PREF_AUTH_USERNAME;
 import static com.github.niqdev.ipcam.settings.SettingsActivity.PREF_IPCAM_URL;
 
-public class IpCamSnapshotActivity extends AppCompatActivity implements OnFrameCapturedListener {
+public class IpCamSnapshotActivity extends AppCompatActivity  {
 
     private static final int TIMEOUT = 5;
     private static final String TAG = "IpCamSnapshotActivity";
-    private static final int REQUEST_CODE = 1;
 
     @BindView(R.id.mjpegViewSnapshot)
     MjpegView mjpegView;
@@ -49,22 +46,19 @@ public class IpCamSnapshotActivity extends AppCompatActivity implements OnFrameC
     ImageView snapshotView;
     @BindView(R.id.record_text)
     TextView timerText;
-    private Bitmap lastPreview = null;
     private Timer timer = new Timer();
     int cnt = 0;
-    private boolean isRecording = false;
-    private FileOutputStream fos;
-    private BufferedOutputStream bos;
+    private MjpegRecordingHandler recordingHandler;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_ipcam_snapshot);
         ButterKnife.bind(this);
-        mjpegView.setOnFrameCapturedListener(this);
+        recordingHandler=new MjpegRecordingHandler(this);
+        mjpegView.setOnFrameCapturedListener(recordingHandler);
         timerText.setText("00:00:00");
     }
-
 
     private String getPreference(String key) {
         return PreferenceManager
@@ -118,46 +112,28 @@ public class IpCamSnapshotActivity extends AppCompatActivity implements OnFrameC
         switch (item.getItemId()) {
             case R.id.action_capture:
                 runOnUiThread(() -> {
-                    if (lastPreview != null) {
+                    if (recordingHandler.getLastBitmap() != null) {
                         snapshotView.setVisibility(View.VISIBLE);
-                        snapshotView.setImageBitmap(lastPreview);
-                        saveBitmapToFile(lastPreview);
+                        snapshotView.setImageBitmap(recordingHandler.getLastBitmap());
+                        recordingHandler.saveBitmapToFile();
                     }
                 });
                 return true;
             case R.id.action_recording:
-                isRecording = !isRecording;
-                item.setIcon(isRecording ? R.drawable.recording : R.drawable.ic_videocam_white_48dp);
-                if (isRecording) {
+                if (!recordingHandler.isRecording()) {
                     startRecording();
                 } else {
                     stopRecording();
                 }
+                item.setIcon(recordingHandler.isRecording() ? R.drawable.recording : R.drawable.ic_videocam_white_48dp);
                 return true;
             default:
                 return super.onOptionsItemSelected(item);
         }
     }
 
-    public void saveBitmapToFile(Bitmap bmp) {
-        FileOutputStream fos = null;
-        BufferedOutputStream bos = null;
-        String imagePath = createJpgFile().getAbsolutePath();
-        try {
-            fos = new FileOutputStream(imagePath);
-            bos = new BufferedOutputStream(fos);
-            ByteArrayOutputStream jpegByteArrayOutputStream = new ByteArrayOutputStream();
-            bmp.compress(Bitmap.CompressFormat.JPEG, 75, jpegByteArrayOutputStream);
-            byte[] jpegByteArray = jpegByteArrayOutputStream.toByteArray();
-            bos.write(jpegByteArray);
-            bos.flush();
-            Toast.makeText(this, "saved image:" + imagePath, Toast.LENGTH_LONG).show();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
     private void stopRecording() {
+        recordingHandler.stopRecording();
         timer.cancel();
         timer.purge();
         timer = null;
@@ -180,75 +156,12 @@ public class IpCamSnapshotActivity extends AppCompatActivity implements OnFrameC
         timer.schedule(timerTask, 0, 1000);
         timerText.setVisibility(View.VISIBLE);
         timerText.setText("00:00:00");
-        try {
-            String mjpegFilePath = createMjpegFile().getAbsolutePath();
-            fos = new FileOutputStream(mjpegFilePath);
-            bos = new BufferedOutputStream(fos);
-            Toast.makeText(this, "start recording, file path is:" + mjpegFilePath, Toast.LENGTH_LONG).show();
-            isRecording = true;
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
-        }
-
+        recordingHandler.startRecording();
     }
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         getMenuInflater().inflate(R.menu.menu_capture, menu);
         return true;
-    }
-
-    @Override
-    public void onFrameCaptured(Bitmap bitmap) {
-        lastPreview = bitmap;
-    }
-
-    @Override
-    public void onFrameCapturedWithHeader(byte[] bitmapData, byte[] header) {
-        if (isRecording) {
-            try {
-                bos.write(header);
-                bos.write(bitmapData);
-                bos.flush();
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-    }
-
-    private File createJpgFile() {
-        Date T = new Date();
-        SimpleDateFormat sdf = new SimpleDateFormat("yyyyMMddHHmmss");
-        String szFileName = "photo-" + sdf.format(T);
-        try {
-            String path = getExternalFilesDir(null).getPath() + "/" + szFileName + ".jpg";
-            File file = new File(path);
-            if (!file.exists()) {
-                file.createNewFile();
-            }
-            Log.d(TAG, "file path is " + file.getAbsolutePath());
-            return file;
-        } catch (IOException e) {
-            Log.v(TAG, e.getMessage());
-        }
-        return null;
-    }
-
-    private File createMjpegFile() {
-        Date T = new Date();
-        SimpleDateFormat sdf = new SimpleDateFormat("yyyyMMddHHmmss");
-        String szFileName = "video-" + sdf.format(T);
-        try {
-            String path = getExternalFilesDir(null).getPath() + "/" + szFileName + ".mjpeg";
-            File file = new File(path);
-            if (!file.exists()) {
-                file.createNewFile();
-            }
-            Log.d(TAG, "file path is " + file.getAbsolutePath());
-            return file;
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return null;
     }
 }
